acctest: Split common configurations into their own source file.
